More stuff n things.
I corrected the degree column. I may want to do something similar for the position column? Those are ordinal, too... ? Hm. Also created the column for the 3 categories of diversity (none, 1, and 2 or more). Then re-ran the correlations. I'm at the chi-squared tests and need to add the normalized proportions for some and also the bonferroni corrections.