I have questions.
In using the exact correlation stat sig threshold, there are a lot more stat sig correlations, and some include items of diversity. I didn't do any additional feature engineering, though... I did do some modeling (but with the feature engineering from the conservative correlation time ... ) and I'm not sure I did them correctly. If I did, there seem to be two good models (one with the binary column as a target and another with the sum+female column as a target). I don't know what this means... And what are the benefits/limitations of using the conservative estimates vs. exact for the correlations?